refactor: transaction decorator 적용
